requirements.txt
torch
torchvision
tqdm
triton
typing_extensions
einops==0.8.0
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe
mpmath==1.3.0
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pillow
PyYAML==6.0.2
sympy==1.13.1
six
segmentation-models-pytorch